requirements.txt
django
djangorestframework
drf_yasg
Pillow
python-decouple
django-storages
boto3
django-filter